More model refactoring
- Decoupled classification task from model
- Introduced a `VisionColumnformer` model that wraps a `Columnformer`
  and adds patching, prediction head, etc. Aims to be mostly consistent
  with timm `VisionTransformer`.
