+++ columnformers/models/__init__.py
@@ -1,3 +1,4 @@
-from . import classification  # noqa
 from . import columnformer_v1  # noqa
+from . import vision_columnformer  # noqa
+from . import vision_transformer  # noqa
 from .registry import create_model, list_models  # noqa
