+++ columnformers/tasks/__init__.py
@@ -1,0 +1,1 @@
+from .classification import ImageClassification  # noqa
